Huge JS change
This change adds in a few statements that refer to all coaches so that I won't refer to each one. In addition, the code allows me to have the links be separate from the teacher so that I can set the links to display separately. This removes the crazy repetitive lines from the JS code. Other than that it might improve performance a small bit but I don't know for sure. Either way, the JS code shouldn't take up 80% of our project anymore.